add student control panel
